% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{paadcr_auth}
\alias{paadcr_auth}
\title{Query a week from PAADC API}
\usage{
paadcr_auth(user, pw)
}
\arguments{
\item{user}{The username}

\item{pw}{The password}
}
\description{
This function will add a username and a password to your
\code{.Renviron} file so it can be called securely without being stored in
your code. After you have installed your credentials, it can be called any time by
typing \code{Sys.getenv("PAADC_USER")} or \code{Sys.getenv("PAADC_PW")}. If you do not have an \code{.Renviron} file, the
function will create one for you.
}
\examples{

\dontrun{

# Just run \code{paadcr_auth()} and the function will take care of the rest.
paadcr_auth()

}
}
