% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{paadcr_week}
\alias{paadcr_week}
\title{Query a week}
\usage{
paadcr_week(week, clean = F)
}
\arguments{
\item{week}{Specify the week as numeric. For example for Week 2:  Jan 11 - 17 2021.}

\item{clean}{clean the dataset}
}
\description{
This function will retrieve data from any given week of the API.
}
\examples{

\dontrun{

# Query week 2:
paadcr_week(2)

}
}
